check that lanes>1 when testing left lane stuff
I was bailing out on one lane highways.  Data seemed okay, but the only
error condition was that left lane was no counts.  So fixed that test to
only carp if there was actually a left lane (lanes>1).

Seems to work okay in Amelia, and so far at least one site with lanes>1
has passed the test and imputed okay, so I think this is a good fix.