fixed plotting bug
The issue was a couple of things, but essentially, I fixed the
aggregation code a little, and fixed the plotting code so that if it
tries to aggregate to an hour but there is nothing there (because there
are no or very few hours without any missings) then drop down to a
smaller aggregate quantity and retry.

Also changed the titling of the plots to say the proper time step (2mins
or 10 mins etc etc up to 1 hour) and changed the number 4 plot to just
show lanes data if there are smaller-than-hour aggregation steps.