Not much I can do for this bug
A bunch of Amelia runs were done with an offset time.  Not much I can do
to fix from within the plotting.

Two options:  redo the imputations, or load the original and imputed
data, compare the starting times, and decide whether or not the
timestamp is off by 8 hours, and if so fix it and re-call the plot step.