hacking to fix speed plot
impute speed by itself, not as weighted speed

seems to work fine so far, but need to test for 2012 all

The issue is that the speed is really the weighted average of speeds,
but if you impute the count and the weighted speed at the same time,
they deviate and you can get crazy speeds.  So I'm not computing the ave
speed by dividing by count prior to imputation, so that the imputation
has the right bounds and does the right thing (keeps speed reasonable).