fix and test for bad file pattern bug
noticed that my file pattern was bad, and that I was processing
detectors for which I did not have data.  For example, have 7005, was
matching for detector 7.

So changed the pattern matcher to work, and also put a filter in the
loader such that if any row does not have the expected detector
station id (detstaid field) matching the requested tams id, then the
row is filtered out.

tests pass