hackity hack
hopefully this is less memory intensive.  I tried to be carefull about
passing the CSV data around, so as to minimize the chance of random
copies.

R does a sort of pass by promise, so apparently if you are careful
about not mutating the object while diving down the stack, you can
avoid copies.  But as soon as stuff changes, R will make a copy.

Anyway, the code is much uglier, but perhaps less chance of copying
data.

Last try in any event

And the tams impute tests pass.